Add navbar and restructure the project.
