+++ src/components/home/Home.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Home = () => (
+  <section>
+    <h2>Welcome to our page!</h2>
+  </section>
+);
+
+export default Home;
