+++ src/components/home/home.css
@@ -1,0 +1,1 @@
+/* stylesheet */
