+++ src/components/navBar/nav.css
@@ -1,0 +1,53 @@
+header {
+  width: 100%;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1.5rem 5.5rem;
+}
+
+header h1 {
+  font-family: var(--ff-righteous);
+  font-size: 2.5rem;
+}
+
+.nav--main li > a::after {
+  content: '';
+  display: block;
+  width: 0%;
+  height: 3px;
+  background-color: var(--clr-dark);
+  animation: draw-out 0.5s cubic-bezier(0.19, 1, 0.22, 1);
+}
+
+.nav--main li > a:hover::after {
+  content: '';
+  display: block;
+  width: 100%;
+  height: 3px;
+  background-color: var(--clr-dark);
+  animation: draw-in 0.5s cubic-bezier(0.19, 1, 0.22, 1);
+}
+
+@keyframes draw-in {
+  0% {
+    width: 0%;
+    opacity: 0;
+  }
+
+  100% {
+    width: 100%;
+    opacity: 1;
+  }
+}
+
+@keyframes draw-out {
+  0% {
+    width: 100%;
+    opacity: 1;
+  }
+
+  100% {
+    width: 0%;
+    opacity: 0;
+  }
+}
