+++ src/components/quote/quote.css
@@ -1,0 +1,1 @@
+/* stylesheet */
