+++ src/index.css
@@ -1,4 +1,5 @@
-@import url('http://fonts.cdnfonts.com/css/poppins');
+@import url('https://fonts.googleapis.com/css2?family=Righteous&display=swap');
+@import url('http://fonts.cdnfonts.com/css/roboto');
 
 :root {
   /* font families */
@@ -8,8 +9,8 @@
   --clr-gray: #858693;
 
   /* font families */
-  --ff-sugo: 'Sugo Pro Classic Trial', sans-serif;
-  --ff-poppins: 'Poppins', sans-serif;
+  --ff-righteous: 'Righteous', cursive;
+  --ff-roboto: 'Roboto', sans-serif;
 }
 
 /* RESET */
@@ -59,7 +60,7 @@ html:focus-within {
 body {
   text-rendering: optimizeSpeed;
   line-height: 1.5;
-  font-family: var(--ff-poppins);
+  font-family: var(--ff-roboto);
   scroll-behavior: smooth;
 }
 
@@ -153,7 +154,7 @@ input[type='search'] {
 }
 
 .nav--main {
-  font-family: var(--ff-sugo);
+  font-family: var(--ff-righteous);
   font-size: 1.5rem;
   font-weight: 300;
   color: var(--clr-dark);
